requirements.txt
mcp>=1.0.0
watchdog>=3.0.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
psutil>=5.9.0
pyyaml>=6.0
aiohttp>=3.8.0
python-dotenv>=1.0.1
requests>=2.31.0
PyQt6>=6.6.0
